requirements.txt
behave-django
django-codemirror2
django-cte
django-mptt
factory-boy
mock
pyhamcrest
